pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rjobs"
version = "0.1.0"
description = "Search and rank remote job opportunities across multiple sources"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "openai>=1.0",
    "pymupdf>=1.24",
]

[project.optional-dependencies]
linkedin = [
    "playwright>=1.40",
]
dev = [
    "pre-commit",
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[project.scripts]
rjobs = "rjobs.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["rjobs"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
